Remove MUP
MUP is used internally at Lempire to deploy Lemverse, but it's quite tied with
their own setup and isn't very practical.

Let's remove it so that it doesn't confuse new users and makes onboarding
simpler.

PS: I'm moving `/public/app` to `/app` as a result, since there is nothing else
in `/public` anymore.