Make the developer onboarding simpler and platform-agnostic
The scripts in the repository give the impression that running the project is
quite complicated, whereas running a Meteor app is actually a simple 3 steps
process:

* Install system dependencies if any
* Run `meteor npm install`
* Run `meteor`

Also, the current scripts do quite a bunch of extra stuff that's not strictly
necessary, and are quite platform dependent.

Let's KISS by removing the scripts and expose the real commands.