Improve discussions analytics
DiscussionAttend and DiscussionEnd events are now thrown also when a
user joins or leaves an existing discussion.

users_attending_count: This property allows to
know how many users were in the discussions before the join/left and
thus allows to filter new discussions/discussions ending if needed.

peer_user_id: The user id of the user joining/leaving the discussion.

The choice of throwing the same events with a users_attending_count rather than new
events (e.g. DiscussionJoin, DiscussionLeft) is because it is easier in most analytics tools
to filter on properties rather than aggregating events.