Add passwordless login
Use Meteor accounts-passwordless package to implement passwordless login.

When the setting "passwordless" is set to true, the password step is skipped
during registration process.

During the login, the user only needs to provide his email and receives a link to be
automatically logged-in.

The startup hook to auto login with query param token is required because the default
accounts-passwordless implementation isn't compatible with the accounts-guest package.