Monti support (#140)
* Add Monti APM support

* Enable Monti APM CPU profiling

* Fix V8 profiling by compiling and running with the same libc

montiapm:profiler depends on v8-profiler-next that compiles node bindings for
the V8 profiler. Compiling against glibc (Debian) and then running against
musl (Alpine) is usually a bad idea and in our case, it seems to make the
profiler segfault. This commit thus fixes a segfault when we try to CPU profile
our app with Monti APM.